Updated helm to the latest version.

+++ elisp/helm/emacs-helm.sh
@@ -23,19 +23,53 @@
 # Run it from this directory.
 
 TMP="/tmp/helm-cfg.el"
-LOADPATH=`dirname $0`
+EMACS=emacs
+
+case $1 in
+    -P)
+        shift 1
+        declare EMACS=$1
+        shift 1
+        ;;
+    -h)
+        echo "Usage: ${0##*/} [-P} Emacs path [-h} help [--] EMACS ARGS"
+        exit 2
+        ;;
+esac
+
+cd $(dirname "$0")
+
+# Check if autoload file exists.
+# It is maybe in a different directory if
+# emacs-helm.sh is a symlink.
+LS=$(ls -l $0 | awk '{print $11}')
+if [ ! -z $LS ]; then
+    AUTO_FILE="$(dirname $LS)/helm-autoloads.el"
+else
+    AUTO_FILE="helm-autoloads.el"
+fi
+if [ ! -e "$AUTO_FILE" ]; then
+    echo No autoloads found, please run make first to generate autoload file
+    exit 2
+fi
+
+
 cat > $TMP <<EOF
 (setq initial-scratch-message (concat initial-scratch-message
 ";;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;\n\
-;; This Emacs is Powered by \`HELM'.\n\
+;; This Emacs is Powered by \`HELM' using\n\
+;; emacs program \"$EMACS\".\n\
 ;; This is a minimal \`helm' configuration to discover \`helm' or debug it.\n\
 ;; You can retrieve this minimal configuration in \"$TMP\" \n\
 ;; Some originals emacs commands have been replaced by own \`helm' commands:\n\n\
 ;; - \`find-file'(C-x C-f)           =>\`helm-find-files'\n\
 ;; - \`occur'(M-s o)                 =>\`helm-occur'\n\
 ;; - \`list-buffers'(C-x C-b)        =>\`helm-buffers-list'\n\
-;; - \`completion-at-point'(M-tab)   =>\`helm-lisp-completion-at-point'\n\n\
+;; - \`completion-at-point'(M-tab)   =>\`helm-lisp-completion-at-point'[1]\n\
+;; - \`dabbrev-expand'(M-/)          =>\`helm-dabbrev'\n\n\
 ;; Some others native emacs commands are \"helmized\" by \`helm-mode'.\n\
+;; [1] Coming with emacs-24.4 \`completion-at-point' is \"helmized\" by \`helm-mode'\n\
+;; which provide helm completion in many other places like \`shell-mode'.\n\
 ;; You will find embeded help for most helm commands with \`C-c ?'.\n\
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;\n\n"))
 
@@ -44,15 +78,18 @@ cat > $TMP <<EOF
                             (menu-bar-lines . 0)
                             (fullscreen . nil)))
 (blink-cursor-mode -1)
-(add-to-list 'load-path (expand-file-name "$LOADPATH"))
+(add-to-list 'load-path (file-name-directory (file-truename "$0")))
 (require 'helm-config)
 (helm-mode 1)
 (define-key global-map [remap find-file] 'helm-find-files)
 (define-key global-map [remap occur] 'helm-occur)
 (define-key global-map [remap list-buffers] 'helm-buffers-list)
-(define-key lisp-interaction-mode-map [remap completion-at-point] 'helm-lisp-completion-at-point)
-(define-key emacs-lisp-mode-map       [remap completion-at-point] 'helm-lisp-completion-at-point)
-(add-hook 'kill-emacs-hook #'(lambda () (delete-file "$TMP")))
-(cd "~/")
+(define-key global-map [remap dabbrev-expand] 'helm-dabbrev)
+(unless (boundp 'completion-in-region-function)
+  (define-key lisp-interaction-mode-map [remap completion-at-point] 'helm-lisp-completion-at-point)
+  (define-key emacs-lisp-mode-map       [remap completion-at-point] 'helm-lisp-completion-at-point))
+(add-hook 'kill-emacs-hook #'(lambda () (and (file-exists-p "$TMP") (delete-file "$TMP"))))
 EOF
-emacs -Q -l $TMP $@
+
+$EMACS -Q -l $TMP $@
+
